Data type for monetary values changed (double -> BigDecimal)

+++ src/main/java/org/td024/entity/Workspace.java
@@ -2,6 +2,8 @@ package org.td024.entity;
 
 import org.td024.enums.WorkspaceType;
 
+import java.math.BigDecimal;
+
 public class Workspace {
     private int id;
 
@@ -9,9 +11,9 @@ public class Workspace {
 
     private WorkspaceType type;
 
-    private double price;
+    private BigDecimal price;
 
-    public Workspace(String name, WorkspaceType type, double price) {
+    public Workspace(String name, WorkspaceType type, BigDecimal price) {
         this.name = name;
 
         this.type = type;
@@ -43,11 +45,11 @@ public class Workspace {
         this.type = type;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 }
